This is a wrapper function for DAISIE::DAISIE_dataprep(). It allows the final DAISIE data structure to be produced from within DAISIEprep. For detailed documentation see the help documentation in the DAISIE package (?DAISIE::DAISIE_dataprep).

This is a wrapper function for DAISIE::DAISIE_dataprep(). It allows the final DAISIE data structure to be produced from within DAISIEprep. For detailed documentation see the help documentation in the DAISIE package (?DAISIE::DAISIE_dataprep).

Usage

create_daisie_data(
  daisie_datatable,
  island_age,
  num_mainland_species,
  num_clade_types = 1,
  list_type2_clades = NA,
  prop_type2_pool = "proportional",
  epss = 1e-05,
  verbose = FALSE
)

Arguments

daisie_datatable

A data frame where each row on the table represents an independent colonisation event. The table has the following four columns: * Clade_name: name of independent colonisation event * Status: One of the following categories: - "Non_endemic": applies to non-endemic species when an approximate colonisation time is known. - "Non_endemic_MaxAge": applies to non-endemic species for cases where colonisation time is unknown. - "Endemic": applies to endemic species or endemic clades when an approximate colonisation time is known. - "Endemic_MaxAge": appies to endemic species or endemic clades where the colonisation time is unknown, or when the user wants to specify an upper bound for colonisation. This could for example apply to endemic species that have recently gone extinct because of anthropogenic causes, and which are not included in the phylogeny ("NA" should be given in the branching times column). It could also apply to insular radiations with long stem branches, for which the time of the first cladogenetic event is known, but the precise time of the colonisation is not. - "Endemic&Non_Endemic": where endemic clade is present and its mainland ancestor has re-colonised. * Missing_species: Number of island species that were not sampled for a particular clade (only applicable for "Endemic" clades). If NA is given in branching times column, this should be equal to the number of species in the clade minus 1. * Branching_times: Stem age of the population/species in the case of "Non_endemic", "Non_endemic_MaxAge", and "Endemic" species with no extant clade relatives on the island. Set "NA" if colonisation time is unknown and no upper bound is known. For "Endemic" cladogenetic species these should be branching times of the radiation, including the stem age of the radiation (colonisaton time estimate).

island_age

Age of the island in appropriate units.

num_mainland_species

The size of the mainland pool, i.e. the number of species that can potentially colonise the island.

num_clade_types

Number of clade types. Default num_clade_types = 1 all species are considered to belong to the same macroevolutionary process. If num_clade_types = 2, there are two types of clades with distinct macroevolutionary processes.

list_type2_clades

If num_clade_types = 2, list_type2_clades specifies the names of the clades that have a distinct macroevolutionary process. The names must match those in the "Clade_name" column of the source data table. If num_clade_types = 1, then list_type2_clades = NA should be specified (default).

prop_type2_pool

Specifies the fraction of potential mainland colonists that have a distinct macroevolutionary process. Applies only if number_clade_types = 2. Default "proportional" sets the fraction to be proportional to the number of clades of distinct macroevolutionary process that have colonised the island. Alternatively, the user can specify a value between 0 and 1 (e.g. if the mainland pool size is 1000 and prop_type2_pool = 0.02 then the number of type 2 species is 20).

epss

Default = 1e-5 should be appropriate in most cases. This value is used to set the maximum age of colonisation of "Non_endemic_MaxAge" and "Endemic_MaxAge" species to an age that is slightly younger than the island for cases when the age provided for that species is older than the island. The new maximum age is then used as an upper bound to integrate over all possible colonisation times.

verbose

Boolean. States if intermediate results should be printed to console. Defaults to TRUE.

Value

DAISIE data list

Examples

if (FALSE) {
island_tbl <- extract_island_specie()
daisie_datatable <- as_daisie_datatable(island_tbl)
daisie_data_list <- create_daisie_data(
  daisie_datatable = daisie_datatable,
  island_age = 1,
  num_mainland_species = 1000,
  num_clade_types = 1,
  list_type2_clades = NA,
  prop_type2_pool = NA,
  epss = 1e-5,
  verbose = FALSE
)
}